ccn-lite-ctrl: fix parsing of UDP string
In the current version the `ccn-lite-ctrl` command tries to send to the
port equal to the first byte of the given IP address when used with
`-u`.

Calling `strtok()` with the first parameter `!= NULL` returns the
*first* token of a tokenized string. Subsequent tokens are retrieved
with a subsequent call of `strtok` with the first parameter being
`NULL`.

This commit sets the result of `strtok()` to `udp` (the IP address
string in the later program) and uses the subsequent call with `NULL`
to parse the UDP port.