ccnl-relay: check if face ifndx matches given ifndx
Not really sure this is correct, but I have the situation where I have
2 `AF_PACKET` interfaces in use with RONR and CCN-lite is only
broadcasting out one of the interfaces. This is happening, because both
interfaces get the same broadcast `sockaddr` down in 471-477, so `sa`
is matching at the fixed line with `f->peer` of the first interface a
broadcast face was created for.

This fixes the problem for me, but where I'm not sure is how other face
types might be affected by this change.